Rename files to match implemented classes
